[ZGC] Add GC ID to the ZGCCycle (#229)
The GC id can be useful when analyzing data and may be valuable to users. It can help order data later or give an indication of how many GC cycles have happened since since the process start or between points in time.

I've plumbed the GC id from the ZGCParser through to the ZGCCycle class. It'll be captured as part of the cycle start regex.