[zgc] Add support for relocation summary
This commit adds support for all the information for gc,reloc stats. This includes Small, Medium and Large page stats, bytes forward and the age table for all the pages, including Eden and Survivor generations.

The Small, Medium and Large table stats are directly accessible on the ZGC cycle, where the page age summaries for Eden and all Survivor generations are included as a variable sized list. As far as I can tell this list can be as long at 14.