Optimize decorator parsing
Profiling showed that a relatively high percentage of CPU time and memory
allocations are coming from decorator parsing. This makes sense, since
this parsing is applied to every unified log line which goes through
gctoolkit.

To speed things up, we:
- Fold the tags regex into the broader decorator regex, avoiding the
  need for a second matcher, and removing the need for an expensive
  negative lookbehind in the tag pattern
- Add a start of line anchor to the decorator regex
- Defer sanitization of tags until getTags() is called for the first time

To reduce unnecessary memory allocations, match groups are retrieved
once and stored in an array.

In testing, these optimizations cut the running time of gctoolkit
roughly in half.